+++ services/k8/ingress.js
@@ -70,7 +70,7 @@ const getHostnameFromRegex = (url) => {
 const getK8sApi = () => {
   const kc = new k8s.KubeConfig();
   kc.loadFromCluster();
-  return k8sApi = kc.makeApiClient(k8s.NetworkingV1beta1Api);
+  return k8sApi = kc.makeApiClient(k8s.NetworkingV1Api);
 }
 
 
@@ -149,8 +149,11 @@ const formatIngressName = (domain) => {
 /**
  * Return the body to create / replace a namespaced ingress through the API
  *
+ * @param ingressName
  * @param domain
- * @returns {{metadata: {name: *, annotations: {"cert-manager.io/cluster-issuer": string, "kubernetes.io/ingress.class": string}}, apiVersions: string, kind: string, spec: {rules: [{host: *, http: {paths: [{path: string, backend: {servicePort: number, serviceName: string}}]}}], tls: [{secretName: *, hosts: [*]}]}}}
+ * @param addWww
+ * @param secretName
+ * @returns {{metadata: {name, annotations: {"nginx.ingress.kubernetes.io/proxy-body-size": string, "nginx.ingress.kubernetes.io/configuration-snippet": string, "nginx.ingress.kubernetes.io/from-to-www-redirect": string, "kubernetes.io/ingress.class": string}, labels: {type: string}}, apiVersions: string, kind: string, spec: {rules: [{host, http: {paths: [{path: string, backend: {service: {port: {number: (number|number)}, name: (string|string)}}, pathType: string}]}}], tls: [{secretName: *, hosts: *[]}]}}}
  */
 const getIngressBody = (ingressName, domain, addWww, secretName) => {
   const domains = [domain];
@@ -159,8 +162,8 @@ const getIngressBody = (ingressName, domain, addWww, secretName) => {
     domains.push('www.' + domain);
   }
 
-  return {
-    apiVersions: 'networking.k8s.io/v1beta1',
+  const ingressBody = {
+    apiVersions: 'networking.k8s.io/v1',
     kind: 'Ingress',
     metadata: {
       labels: {
@@ -168,7 +171,6 @@ const getIngressBody = (ingressName, domain, addWww, secretName) => {
       },
       name: ingressName,
       annotations: {
-        'cert-manager.io/cluster-issuer': 'openstad-letsencrypt-prod', // Todo: make this configurable
         'kubernetes.io/ingress.class': 'nginx',
         'nginx.ingress.kubernetes.io/from-to-www-redirect': "true",
         'nginx.ingress.kubernetes.io/proxy-body-size': '128m',
@@ -183,11 +185,16 @@ more_set_headers "Referrer-Policy: same-origin";`
         host: domain,
         http: {
           paths: [{
+            path: '/',
+            pathType: 'Prefix',
             backend: {
-              serviceName: 'openstad-frontend', // Todo: make this configurable
-              servicePort: 4444 // Todo: make this configurable
-            },
-            path: '/'
+              service: {
+                name: process.env.KUBERNETES_FRONTEND_SERVICE_NAME || 'openstad-frontend',
+                port: {
+                  number: process.env.KUBERNETES_FRONTEND_SERVICE_PORT ? parseInt(process.env.KUBERNETES_FRONTEND_SERVICE_PORT) : 4444
+                }
+              }
+            }
           }]
         }
       }],
@@ -197,6 +204,12 @@ more_set_headers "Referrer-Policy: same-origin";`
       }]
     }
   }
+  
+  if (process.env.KUBERNETES_CLUSTER_ISSUER_ENABLED) {
+    ingressBody.metadata.annotations['cert-manager.io/cluster-issuer'] = process.env.KUBERNETES_CLUSTER_ISSUER || 'openstad-letsencrypt-prod';
+  }
+  
+  return ingressBody;
 };
 const getAll = async () => {
   let response = await getK8sApi().listNamespacedIngress(process.env.KUBERNETES_NAMESPACE);
@@ -333,7 +346,9 @@ exports.ensureIngressForAllDomains = async () => {
 
   });
 
-  const systemIngresses = ['openstad-admin', "openstad-frontend", "openstad-image", "openstad-api", "openstad-auth"];
+  const ingressPrefix = (process.env.KUBERNETES_NAMESPACE ? process.env.KUBERNETES_NAMESPACE : 'openstad');
+  
+  const systemIngresses = [`${ingressPrefix}-admin`, `${ingressPrefix}-frontend`, `${ingressPrefix}-image`, `${ingressPrefix}-api`, `${ingressPrefix}-auth`];
 
   // filter all domains present
   let domainsToDelete = Object.keys(domainsInIngress).filter((domainInIngress) => {
@@ -506,7 +521,3 @@ const processIngressForDomain = async (domain, sites, ingressName) => {
  * @returns {Promise<{response: http.IncomingMessage; body: NetworkingV1beta1Ingress}>}
  */
 exports.add = add;
-
-
-
-